backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy==1.24.3
scipy==1.11.4
opencv-python-headless==4.8.1.78
imutils==0.5.4
python-magic==0.4.27
Pillow==10.0.1
pydantic==2.5.0
psutil==5.9.6
